requirements.txt
numpy>=1.22.0
pandas>=1.5.0
scikit-learn>=1.1.3
tensorflow>=2.12.0
google-cloud-bigquery>=3.10.0
google-cloud-storage==2.14.0
google-api-core==2.8.2
googleapis-common-protos==1.56.4
matplotlib>=3.5.0
seaborn>=0.12.0
flask>=2.2.0
gunicorn>=20.0.4
ipykernel
plotly
librosa
google-cloud-bigquery
matplotlib
nbformat>=4.2.0
mne
